warnings 2

+++ gpu_plasma.h
@@ -2021,8 +2021,8 @@ int memory_monitor(std::string legend,int nt)
 
 	cudaError_t err = cudaMemGetInfo(&m_free,&m_total);
 
-	sysinfo(&info);
-	fprintf(f,"step %10d %50s GPU memory total %10d free %10d free CPU memory %10u \n",nt,legend.c_str(),m_total/1024/1024,m_free/1024/1024,info.freeram/1024/1024);
+	sysinfo(&info);                                                                //  1   2              3                 4                5
+	fprintf(f,"step %10d %50s GPU memory total %10d free %10d free CPU memory %10u \n",nt,legend.c_str(),((int)m_total)/1024/1024,((int)m_free)/1024/1024,((int)info.freeram)/1024/1024);
 
 }
